+++ packages/webview_flutter/webview_flutter_android/pigeons/android_webview.dart
@@ -86,6 +86,17 @@ class WebViewPoint {
   int y;
 }
 
+class GeoPermissionsHandleResult {
+  GeoPermissionsHandleResult({
+    this.origin = '',
+    this.isAllow = false,
+    this.isRetain = false,
+  });
+  String origin;
+  bool isAllow;
+  bool isRetain;
+}
+
 /// Handles methods calls to the native Java Object class.
 ///
 /// Also handles calls to remove the reference to an instance with `dispose`.
@@ -328,6 +339,12 @@ abstract class WebChromeClientFlutterApi {
     int instanceId,
     int webViewInstanceId,
     int paramsInstanceId,
+  );
+
+  @async
+  GeoPermissionsHandleResult onGeolocationPermissionsShowPrompt(
+      int instanceId,
+      String origin,
   );
 }