+++ packages/webview_flutter/webview_flutter_android/test/android_navigation_delegate_test.dart
@@ -480,6 +480,7 @@ class CapturingWebChromeClient extends android_webview.WebChromeClient {
   CapturingWebChromeClient({
     super.onProgressChanged,
     super.onShowFileChooser,
+    super.onGeolocationPermissionsShowPrompt,
   }) : super.detached() {
     lastCreatedDelegate = this;
   }