+++ packages/webview_flutter/webview_flutter_android/test/android_webview_controller_test.mocks.dart
@@ -341,6 +341,16 @@ class MockAndroidNavigationDelegate extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
+  @override
+  _i9.Future<void> setOnUrlChange(_i3.UrlChangeCallback? onUrlChange) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #setOnUrlChange,
+          [onUrlChange],
+        ),
+        returnValue: _i9.Future<void>.value(),
+        returnValueForMissingStub: _i9.Future<void>.value(),
+      ) as _i9.Future<void>);
 }
 
 /// A class which mocks [AndroidWebViewController].
@@ -674,6 +684,18 @@ class MockAndroidWebViewController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
+  @override
+  _i9.Future<void> setOnGeolocationPermissionsShowPrompt(
+          _i9.Future<_i8.GeolocationPermissionsResult> Function(String)?
+              onGeolocationPermissionsShowPrompt) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #setOnGeolocationPermissionsShowPrompt,
+          [onGeolocationPermissionsShowPrompt],
+        ),
+        returnValue: _i9.Future<void>.value(),
+        returnValueForMissingStub: _i9.Future<void>.value(),
+      ) as _i9.Future<void>);
 }
 
 /// A class which mocks [AndroidWebViewProxy].
@@ -698,6 +720,8 @@ class MockAndroidWebViewProxy extends _i1.Mock
           ) as _i2.WebView Function({required bool useHybridComposition}));
   @override
   _i2.WebChromeClient Function({
+    _i9.Future<_i2.GeoPermissionsHandleResultProxy> Function(String)?
+        onGeolocationPermissionsShowPrompt,
     void Function(
       _i2.WebView,
       int,
@@ -709,6 +733,8 @@ class MockAndroidWebViewProxy extends _i1.Mock
   }) get createAndroidWebChromeClient => (super.noSuchMethod(
         Invocation.getter(#createAndroidWebChromeClient),
         returnValue: ({
+          _i9.Future<_i2.GeoPermissionsHandleResultProxy> Function(String)?
+              onGeolocationPermissionsShowPrompt,
           void Function(
             _i2.WebView,
             int,
@@ -723,6 +749,8 @@ class MockAndroidWebViewProxy extends _i1.Mock
           Invocation.getter(#createAndroidWebChromeClient),
         ),
         returnValueForMissingStub: ({
+          _i9.Future<_i2.GeoPermissionsHandleResultProxy> Function(String)?
+              onGeolocationPermissionsShowPrompt,
           void Function(
             _i2.WebView,
             int,
@@ -737,6 +765,8 @@ class MockAndroidWebViewProxy extends _i1.Mock
           Invocation.getter(#createAndroidWebChromeClient),
         ),
       ) as _i2.WebChromeClient Function({
+        _i9.Future<_i2.GeoPermissionsHandleResultProxy> Function(String)?
+            onGeolocationPermissionsShowPrompt,
         void Function(
           _i2.WebView,
           int,